manage toctrees
large toctrees slow down the build process and blow up the page size. In the remove_from_toctrees list, we can tightly control the depth that we want to be displayed in the sidebar.

Currently, it is set to class level. By removing a ".*", in a line, we can lower it to module level.